Migrate rest of global schedule queries endpoints
