+++ server/service/global_schedule_test.go
@@ -23,6 +23,15 @@ func TestGlobalScheduleAuth(t *testing.T) {
 	ds.NewScheduledQueryFunc = func(ctx context.Context, sq *fleet.ScheduledQuery, opts ...fleet.OptionalArg) (*fleet.ScheduledQuery, error) {
 		return sq, nil
 	}
+	ds.ScheduledQueryFunc = func(ctx context.Context, id uint) (*fleet.ScheduledQuery, error) {
+		return &fleet.ScheduledQuery{}, nil
+	}
+	ds.SaveScheduledQueryFunc = func(ctx context.Context, sq *fleet.ScheduledQuery) (*fleet.ScheduledQuery, error) {
+		return sq, nil
+	}
+	ds.DeleteScheduledQueryFunc = func(ctx context.Context, id uint) error {
+		return nil
+	}
 
 	var testCases = []struct {
 		name            string
@@ -75,6 +84,12 @@ func TestGlobalScheduleAuth(t *testing.T) {
 			checkAuthErr(t, tt.shouldFailRead, err)
 
 			_, err = svc.GlobalScheduleQuery(ctx, &fleet.ScheduledQuery{Name: "query", QueryName: "query"})
+			checkAuthErr(t, tt.shouldFailWrite, err)
+
+			_, err = svc.ModifyGlobalScheduledQueries(ctx, 1, fleet.ScheduledQueryPayload{})
+			checkAuthErr(t, tt.shouldFailWrite, err)
+
+			err = svc.DeleteGlobalScheduledQueries(ctx, 1)
 			checkAuthErr(t, tt.shouldFailWrite, err)
 		})
 	}
